# Replay parsing performance improvements