--- a/data/src/main/java/com/faforever/commons/replay/QtCompress.java
+++ b/data/src/main/java/com/faforever/commons/replay/QtCompress.java
@@ -3,11 +3,8 @@ package com.faforever.commons.replay;
 import java.io.ByteArrayOutputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.zip.Deflater;
-import java.util.zip.DeflaterOutputStream;
-import java.util.zip.Inflater;
-import java.util.zip.InflaterOutputStream;
+import java.nio.ByteBuffer;
+import java.util.zip.*;
 
 /**
  * Utility class that compresses and uncompresses bytes like QT's <a href="http://doc.qt.io/qt-5/qbytearray.html">QByteArray</a>.
@@ -22,17 +19,19 @@ public final class QtCompress {
    * Compresses the specified bytes like <a href="http://doc.qt.io/qt-5/qbytearray.html#qCompress">QByteArray.qCompress()</a>
    * does.
    */
-  public static byte[] qUncompress(byte[] bytes) throws IOException {
+  public static ByteBuffer qUncompress(ByteBuffer inputBuffer) {
     Inflater inflater = new Inflater();
-    inflater.setInput(Arrays.copyOfRange(bytes, 4, bytes.length));
+    final int uncompressedLength = inputBuffer.getInt();
+    inflater.setInput(inputBuffer);
+    final var outputBytes = new byte[uncompressedLength];
 
-    ByteArrayOutputStream byteArray = new ByteArrayOutputStream();
-
-    try (InflaterOutputStream inflaterOutputStream = new InflaterOutputStream(byteArray, inflater)) {
-      inflaterOutputStream.flush();
+    try {
+      inflater.inflate(outputBytes);
+    } catch (DataFormatException e) {
+      throw new RuntimeException(e);
     }
 
-    return byteArray.toByteArray();
+    return ByteBuffer.wrap(outputBytes);
   }
 
   /**