--- a/data/src/main/java/com/faforever/commons/replay/body/ReplayBodyTokenizer.java
+++ b/data/src/main/java/com/faforever/commons/replay/body/ReplayBodyTokenizer.java
@@ -1,11 +1,10 @@
 package com.faforever.commons.replay.body;
 
-import com.google.common.io.LittleEndianDataInputStream;
+import com.faforever.commons.replay.shared.LoadUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.Contract;
-import org.jetbrains.annotations.NotNull;
 
-import java.io.IOException;
+import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,15 +14,15 @@ public class ReplayBodyTokenizer {
   private static final int TOKEN_HEADER_LENGTH = 3;
 
   @Contract(pure = true)
-  public static List<ReplayBodyToken> tokenize(@NotNull LittleEndianDataInputStream dataStream) throws IOException {
+  public static List<ReplayBodyToken> tokenize(ByteBuffer buffer) {
     ArrayList<ReplayBodyToken> tokens = new ArrayList<>();
-    while (dataStream.available() > 0) {
-      int tokenId = dataStream.readUnsignedByte();
-      int tokenLength = dataStream.readUnsignedShort();
+    final int length = buffer.limit();
+    while (buffer.position() < length) {
+      final int tokenId = LoadUtils.getUnsignedByte(buffer);
+      final int tokenLength = buffer.getShort() - TOKEN_HEADER_LENGTH;
 
-      byte[] tokenContent = dataStream.readNBytes(tokenLength - TOKEN_HEADER_LENGTH);
-
-      tokens.add(new ReplayBodyToken(ReplayBodyToken.TokenId.values()[tokenId], tokenLength, tokenContent));
+      buffer.position(buffer.position() + tokenLength);
+      tokens.add(new ReplayBodyToken(ReplayBodyToken.TokenId.values()[tokenId], buffer.position()));
     }
 
     return tokens;