… data/src/main/java/com/faforever/commons/replay/body/ReplayBodyParser.java
… data/src/main/java/com/faforever/commons/replay/header/ReplayHeaderParser.java
… data/src/main/java/com/faforever/commons/replay/shared/LoadUtils.java