### Description

Switched parser logic to ByteBuffer from LittleEndianDataInputStream for more direct and performant access to the originally read replay bytes and easier reusability during parsing

Eliminated double reading of stream and multiple smaller tokenContent stream creation for tokenization and parsing by not preparing separate tokenContent byte chucks and reusing the originally created ByteBuffer

Replaced tokenContent and tokenSize with token limit, which stores the read position limit in the ByteBuffer for the corresponding token

Added buffer limiting to parsing to preserve robustness coming from tokenization

Improved performance of readString methods by replacing the ByteArrayOutputStream and double accessing the ByteBuffer for string ending search and extracting string bytes

Implemented getUnsignedByte util for ByteBuffer

### Measurements
I measured the performance of the parser with the help of VisualVM and the ReplayLoaderTableParserTest test class. CPU: AMD Ryzen 9 5950x. I parsed all my local replays (822 at the time of measurements) with the following code:
```
    sleep(10000);

    File folder = new File("c:/Users/Devastator/Downloads/replays/");
    File[] listOfFiles = folder.listFiles();

    var count = new AtomicInteger();
    for (final var filePath : listOfFiles) {
      System.out.println(count.incrementAndGet());
      new ReplayDataParser(filePath.toPath(), objectMapper);
    }
```

Except for the last parallel test:
```
    sleep(10000);

    File folder = new File("c:/Users/Devastator/Downloads/replays/");
    File[] listOfFiles = folder.listFiles();

    var count = new AtomicInteger();

    ExecutorService EXEC = Executors.newFixedThreadPool(4);
    List<Callable<ReplayDataParser>> tasks = new ArrayList<Callable<ReplayDataParser>>();

    for (final var filePath : listOfFiles) {
      Callable<ReplayDataParser> c = new Callable<ReplayDataParser>() {
        @Override
        public ReplayDataParser call() throws Exception {
          System.out.println(count.incrementAndGet());
          return new ReplayDataParser(filePath.toPath(), objectMapper);
        }
      };
      tasks.add(c);
    }
    List<Future<ReplayDataParser>> results = EXEC.invokeAll(tasks);
```
The sleep call is necessary for having time to attach the profiler to the process and start sampling.

See attached measurements.
[Common Replay Parser Profiling.zip](https://github.com/user-attachments/files/22199360/Common.Replay.Parser.Profiling.zip)

Baseline
**Original.nps:** com.faforever.commons.replay.ReplayDataParser.<init> ()	32 699 ms

Replaced LittleEndianDataInputStream with ByteBuffer just in the tokenizer and used `System.arraycopy()` to extract `TokenContent` bytes
**ByteBuffer tokenizer.nps:** com.faforever.commons.replay.ReplayDataParser.<init> ()	27 798 ms

Used `ByteBuffer `instead of the `LittleEndianDataInputStream ` everywhere throughout the parsing process.
**ByteBuffer full.nps:** com.faforever.commons.replay.ReplayDataParser.<init> ()	22 499 ms

Reused the initially created `ByteBuffer` everywhere to avoid buffer creation for individual tokens.
**ByteBuffer full final.nps:** com.faforever.commons.replay.ReplayDataParser.<init> ()	17 898 ms

Used this measurement to determine how parallelization of different replay parsing is scaling. By not limiting the number of  used threads I ran into out of memory exceptions due to heap constraints, so I elected to limit them according to my measurements and saw the best performance with a limit of 4. Up to 4 threads there were measurable performance improvements, however more than 4 threads made the parsing actually slower.
**ByteBuffer full final parallel.nps:** com.faforever.commons.replay.ReplayLoaderTableParserTest.ExecuteParallel ()	13 299 ms