@Brutus5000 2025-09-10
This is a very interesting contribution and I was not aware of ByteBuffers or even its capability to go LittleEndian. I always wondered why we need Guava to do such a basic thing, well we were using the wrong thing as it seems.

I think the benefits are worth breaking the compatibility here. But we also need to adapt the moderator client later on.

So if we are breaking compatibility anyway we should go all in 😛  Since you are investigating and measuring I have a request: can you check if this also makes sense to appl to loading the data? There we still copy around data and that could be avoided.

This is a suggestion of ChatGPT, I didn't check it

```java
private void readReplayData(Path replayFile) throws IOException, CompressorException {
    // Read everything into a single byte[]
    byte[] allReplayData = Files.readAllBytes(replayFile);
    ByteBuffer buffer = ByteBuffer.wrap(allReplayData);

    // Find header end (still using your helper, which works on byte[])
    int headerEnd = findReplayHeaderEnd(allReplayData);

    // Parse metadata directly from a slice (no Arrays.copyOf)
    ByteBuffer headerSlice = buffer.duplicate();
    headerSlice.limit(headerEnd); // [0, headerEnd)
    metadata = objectMapper.readValue(
        new String(headerSlice.array(), headerSlice.position(), headerSlice.remaining(), StandardCharsets.UTF_8),
        ReplayMetadata.class
    );

    // Slice for the compressed data (no Arrays.copyOfRange)
    ByteBuffer dataSlice = buffer.duplicate();
    dataSlice.position(headerEnd + 1);
    dataSlice = dataSlice.slice(); // starts at headerEnd+1

    // If decompress() can be updated to accept ByteBuffer, pass it directly
    // Otherwise, fall back to array without copying
    if (decompressAcceptsByteBuffer()) {
        data = decompress(dataSlice, metadata);
    } else {
        data = decompress(dataSlice.array(), dataSlice.arrayOffset() + dataSlice.position(), dataSlice.remaining(), metadata);
    }
}
```