@Rogerstalker 2025-09-13
> Since this is a common library and other people might use it, I would rather keep `commandsPerMinuteByPlayer`. Is there any performance reason to remove it?

You can check In the **original.nps** measurement: java.util.HashMap.computeIfAbsent ()	899 ms. Initially I wanted to improve its performance, by using an array instead of the outer Map. This would be possible, because we already know the number of players, that is why I made `playerIdsByName` a class member. But then I saw that it is a private member and it is never used aside from adding data to it, so I thought it might be some now unused, leftover code and removed it.