@Rogerstalker 2025-09-13
Ah I see it had a getter. I tried to check if it was used elsewhere: https://github.com/search?q=org%3AFAForever%20commandsPerMinuteByPlayer&type=code. I don't think it was. It had a performance impact, because was called a lot. We could add it back, but I don't see anyone using it and even if I could improve it, there would still be some impact.