@Brutus5000 2025-09-13
You are right. Should this ever be needed, we'd need to allow passing the events to a consumer passed as a constructor.